+++ src/api/controllers/gallery/list.gallery.js
@@ -1,0 +1,23 @@
+const fs = require('fs');
+const errorHandler = require('../../utils/errorHandler');
+
+module.exports = (app) => {
+  app.get('/gallery', async (req, res) => {
+    try {
+      const location = 'src/gallery';
+      const files = fs.readdirSync(location);
+
+      // Filters and absolute path (without the base url) and removes /src
+      const photos = files
+        .filter((file) => file.match(/^.*\.jpe?g$/))
+        .map((file) => `/${location}/${file}`.replace('/src', ''));
+
+      return res
+        .status(200)
+        .json(photos)
+        .end();
+    } catch (err) {
+      return errorHandler(err, res);
+    }
+  });
+};
