--- src/api/controllers/uploads/get.gallery.js
@@ -1,21 +1,0 @@
-const fs = require('fs');
-const errorHandler = require('../../utils/errorHandler');
-
-module.exports = (app) => {
-  app.get('/uploads/gallery', async (req, res) => {
-    try {
-      const location = 'uploads/gallery';
-      const files = fs.readdirSync(location);
-
-      // Filters and absolute path (without the base url)
-      const photos = files.filter((file) => file.match(/^.*\.jpe?g$/)).map((file) => `/${location}/${file}`);
-
-      return res
-        .status(200)
-        .json(photos)
-        .end();
-    } catch (err) {
-      return errorHandler(err);
-    }
-  });
-};
