+++ src/gallery/.gitignore
