+++ src/images/.gitignore
@@ -1,3 +1,2 @@
 *
 !.gitignore
-!gallery/
